Remove Skip Condition when subject is a Class
Currently, there's a skip condition that checks if the given subject is
a class. This introduces unncessary concern to the ingress library on
handling the subject. As a result this introduces an undesired behaviour
where when a user pass a Class as a given subject, it will always return
true unknowingly that the conditional lambda is never evaluated.

This commit refactors the code to make Ingress a generic and simple
permission checks that will evaluate the conditional lambda regardless
if it's a Class or an Object. The user is responsible for defining the
lambda correctly and handle the subject accordingly.

To make the interface easy, on the next commit, we will introduce 2
conditional lambdas:
* `if_subject_is_a_class`
* `if_subject_is_an_instance`

These lambdas will provide the convenient interface to evaluate
conditional lambda when the given subject is a Class or an Instance.